% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_biodiv.R
\name{pd_biodiv}
\alias{pd_biodiv}
\title{Gather biodiversity records}
\usage{
pd_biodiv(x, db = "gbif", ...)
}
\arguments{
\item{tree}{a phylogeny represented as a newick string}
}
\description{
Gather biodiversity records
}
\examples{
\dontrun{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_phy_tax(tree = st)
res <- pd_tax_hier(x)
res <- pd_query(res, A ~ B)
pd_biodiv(res)
}
}
