% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_meta.R
\name{pd_meta}
\alias{pd_meta}
\title{Metadata repoter}
\usage{
pd_meta(x = last_obj(), file = NULL)
}
\arguments{
\item{x}{an object of class `phylodiv`. defaults to using the last
phylodiv object in the R session}

\item{file}{if `NULL`, metadata report returned. if a file given
we write the metadata summary to the file path as JSON}
}
\value{
- `pd_meta`: a named list of class phylodiv_meta if `is.null(file) == TRUE`, 
or `NULL` if `is.null(file) != TRUE`
- `last_obj`: get the last phylodiv object in the session
}
\description{
Metadata repoter
}
\examples{
\dontrun{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(tree = st)
x
pd_meta()
pd_meta(file = (f <- tempfile()))
readLines(f)
jsonlite::fromJSON(f, FALSE)

# get last phylodiv object
last_obj()
}
}
