% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_phy_tax.R
\name{pd_phy_tax}
\alias{pd_phy_tax}
\title{Gather taxonomic names}
\usage{
pd_phy_tax(tree)
}
\arguments{
\item{tree}{a phylogeny represented as a newick string}
}
\description{
Gather taxonomic names
}
\examples{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
pd_phy_tax(tree = st)
}
