% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_query.R
\name{pd_query}
\alias{pd_query}
\title{Construct phylogenetically based query}
\usage{
pd_query(tree, x, y)
}
\arguments{
\item{x}{a `phylodiv` object}
}
\description{
Construct phylogenetically based query
}
\examples{
\dontrun{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]

# A compared to sister clade
pd_query(st, "Eptesicus bobrinskoi", "Eptesicus bottae")
# pd_query(st, "Eptesicus bobrinskoi", "Eptesicus bar")

# A compared to B
pd_query(st, A ~ B)

# A compared to B
pd_query(st, A ~ B)
}
}
