% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_read.R
\name{pd_read}
\alias{pd_read}
\title{Read a tree to a phylodiv object}
\usage{
pd_read(tree)
}
\arguments{
\item{tree}{a phylogeny represented as a newick string or a
`phylo` object, or a \pkg{tidytree} `tbl_tree` object}
}
\value{
an object of class `phylodiv`, a named list with slots:

- tree: phylo object
- tips: tip labels
- nodes: node labels
}
\description{
Read a tree to a phylodiv object
}
\examples{
# character
x <- "((Strix_aluco:4.2,Asio_otus:4.2):3.1,Athene_noctua:7.3);"
tree <- ape::read.tree(text = x)
z <- pd_read(tree = tree)
z

# phylo
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(tree = st)
x
x$tree
x$tips
x$nodes

# tbl_tree
library(tidytree)
library(ape)
set.seed(2017)
tree <- rtree(4)
x <- as_tibble(tree)
pd_read(x)
}
