% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_tax.R
\name{pd_tax}
\alias{pd_tax}
\title{Gather taxonomic names}
\usage{
pd_tax(x)
}
\arguments{
\item{x}{an object of class `phylodiv`}
}
\value{
an object of class `phylodiv`, a named list with slots:

- tree: phylo object
- tips: tip labels
- nodes: node labels
}
\description{
Gather taxonomic names
}
\examples{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(tree = st)
x <- pd_tax(x)
x
x$tree
x$tips
x$nodes
}
