% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_taxa.R
\name{pd_taxa}
\alias{pd_taxa}
\title{Gather taxonomic hierarchies}
\usage{
pd_taxa(x, db = "ncbi", drop_no_data = TRUE, ...)
}
\arguments{
\item{x}{an object of class `phylodiv`}

\item{db}{(character) the database to use to get 
taxonomic information. only option is "ncbi" for now}

\item{drop_no_data}{(logical) drop tips w/o taxonomic 
hierarchy data. default: `TRUE`}

\item{...}{ignored}
}
\value{
an object of class `phylodiv`, a named list with slots:

- tree
- tips
- nodes
- hierarchies
}
\description{
Gather taxonomic hierarchies
}
\section{taxonomic data}{

Uses \pkg{taxizedb} internally
}

\examples{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
(x <- pd_read(st))
(res <- pd_taxa(x))
res$trees
res$trees[[1]]$hierarchies
res$fetch_hierarchies()
}
