% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_vis.R
\name{pd_vis}
\alias{pd_vis}
\title{Visualize biodiversity records}
\usage{
pd_vis(x, type = "facet", ...)
}
\arguments{
\item{x}{an object of class `phylodiv`}

\item{type}{visualization type, one of count, 
facet, raster}

\item{...}{further arguments passed on to [choroplethr::country_choropleth()] 
(type=facet), [rgbif::map_fetch()] (type=raster)}
}
\value{
a plot
}
\description{
Visualize biodiversity records
}
\examples{
\dontrun{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(st)
res <- pd_taxa(x)
spp <- c("Eptesicus serotinus", "Eptesicus fuscus", 
  "Eptesicus furinalis", "Eptesicus brasiliensis")
res <- pd_query(res, spp)
# skip pd_biodiv to go directly to pd_vis
pd_vis(res, "raster")

# export 
png("myplot.png")
pd_vis(x)
dev.off()


# facet counts on map by category, e.g., country
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(st)
res <- pd_taxa(x)
spp <- c("Eptesicus serotinus", "Eptesicus fuscus", 
  "Eptesicus furinalis", "Eptesicus brasiliensis")
res <- pd_query(res, spp)
biodiv <- pd_biodiv(res, type = 'facet', by = "country")
biodiv
biodiv$data
pd_vis(biodiv, type = "facet")


# counts plotted on tree
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(st)
res <- pd_taxa(x)
spp <- c("Eptesicus serotinus", "Eptesicus fuscus", 
  "Eptesicus furinalis", "Eptesicus brasiliensis")
res <- pd_query(res, spp)
counts <- pd_biodiv(res, type = 'count')
pd_vis(counts, type = "count")

res <- pd_query(res, res$tree$tip.label)
counts <- pd_biodiv(res, type = 'count')
pd_vis(counts, type = "count")
}
}
