% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_viz.R
\name{pd_viz}
\alias{pd_viz}
\title{Visualize biodiversity records}
\usage{
pd_viz(x)
}
\arguments{
\item{x}{results}
}
\description{
Visualize biodiversity records
}
\examples{
\dontrun{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_phy_tax(tree = st)
res <- pd_tax_hier(x)
res <- pd_query(res, A ~ B)
x <- pd_biodiv(res)
pd_viz(x)

# export 
# pd_save_viz(pd_viz(x))
}
}
