% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv-package.R
\docType{package}
\name{phylodiv-package}
\alias{phylodiv-package}
\alias{phylodiv}
\title{Phylogenetic Biodiversity Query Tools}
\description{
Phylogenetic Biodiversity Query Tools
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
