% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloDiv.R
\docType{data}
\name{PhyloDiv}
\alias{PhyloDiv}
\title{PhyloDiv and PhyloDivOne classes}
\description{
PhyloDiv and PhyloDivOne classes
}
\details{
**Methods**
  \describe{
    \item{`add_tree(tree)`}{
      add a tree
    }
    \item{`fetch_hierarchies()`}{
      fetch hierarchy data for all trees
    }
  }

xxx:
}
\examples{
\dontrun{
library(ape)
set.seed(132434)
tree <- rcoal(20)
pd1 <- PhyloDivOne$new(tree)
pd1
pd1$to_list()
pd1$to_json()

x <- "((Strix_aluco:4.2,Asio_otus:4.2):3.1,Athene_noctua:7.3);"
tree <- read.tree(text = x)

# xxx
x <- PhyloDiv$new(pd1, pd1, pd1)
x
x$to_list()
x$to_json()
x$to_json(pretty = TRUE)
}
}
\keyword{internal}
