% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv-package.R
\docType{data}
\name{phylodiv_isocodes}
\alias{phylodiv_isocodes}
\title{Isocodes}
\format{A data frame with 249 rows and 2 variables:

- code: 2 letter country code
- name: country name}
\description{
Isocodes
}
\keyword{data}
