% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_query.R
\name{pd_query}
\alias{pd_query}
\title{Construct phylogenetically based query}
\usage{
pd_query(x, query)
}
\arguments{
\item{x}{an object of class `phylodiv`}

\item{query}{query statement}
}
\description{
Construct phylogenetically based query
}
\section{Use cases}{


Taxonomy based queries (leverages \pkg{taxa}), e.g:
- taxa A vs. taxa B (bypass the tree in this case, but do make 
sure the taxa are in the tree?)
- taxa A, taxa B, taxa C, etc. (bypass the tree in this case, but 
do make sure the taxa are in the tree?)

Phylogeny based queries (no solution for this yet):
- all taxa downstream from A (use tree to get clade from A)
- all taxa upstream from A (use tree to get higher taxa above A)
- all taxa not in genus B (use tree to get all tips not in genus B;
in other words, drop clade B)
}

\examples{
\dontrun{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
st <- ape::subtrees(chiroptera)[[2]]
x <- pd_read(st)
x <- pd_tax(x)
res <- pd_tax_hier(x)

# with a taxmap
res
res$hierarchies
res$taxmap
pd_query(res, startsWith(taxon_names, "Pteropus"))

# with metacoder::parse_phylo()
library(ape)
data(bird.orders)
x <- pd_read(bird.orders)
x <- pd_tax(x)
res <- pd_tax_hier2(x) 
res$taxmap
res <- pd_query(res, startsWith(taxon_names, "C"))


# A compared to sister clade
spp <- c("Eptesicus serotinus", "Eptesicus fuscus", 
  "Eptesicus furinalis", "Eptesicus brasiliensis")
pd_query(res, spp)
# pd_query(res, "Eptesicus bobrinskoi", "Eptesicus bar")

# A compared to B
# pd_query(res, A ~ B)

# A compared to B
# pd_query(res, A ~ B)
}
}
