% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_read.R
\name{pd_read}
\alias{pd_read}
\title{Read a tree to a phylodiv object}
\usage{
pd_read(tree)
}
\arguments{
\item{tree}{a phylogeny represented as a newick string or a
`phylo` object}
}
\value{
an object of class `phylodiv`, a named list with slots:

- tree: phylo object
- tips: tip labels
- nodes: node labels
}
\description{
Read a tree to a phylodiv object
}
\examples{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_read(tree = st)
x
x$tree
x$tips
x$nodes
}
