% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_tax_hier.R
\name{pd_tax_hier}
\alias{pd_tax_hier}
\title{Gather taxonomic hierarchies}
\usage{
pd_tax_hier(x, db = "ncbi", ...)
}
\arguments{
\item{tree}{a phylogeny represented as a newick string}
}
\description{
Gather taxonomic hierarchies
}
\examples{
library(ape)
data(chiroptera)
st <- ape::subtrees(chiroptera)[[393]]
x <- pd_phy_tax(tree = st)
res <- pd_tax_hier(x)
res
res$hierarchies
}
