% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloDiv.R
\name{PhyloDiv}
\alias{PhyloDiv}
\title{PhyloDiv and PhyloDivOne classes}
\description{
PhyloDiv and PhyloDivOne classes

PhyloDiv and PhyloDivOne classes
}
\examples{
\dontrun{
library(ape)
set.seed(132434)
tree <- rcoal(20)
pd1 <- PhyloDivOne$new(tree)
pd1
pd1$to_list()
pd1$to_json()

x <- "((Strix_aluco:4.2,Asio_otus:4.2):3.1,Athene_noctua:7.3);"
tree <- read.tree(text = x)

# xxx
x <- PhyloDiv$new(pd1, pd1, pd1)
x
x$to_list()
x$to_json()
x$to_json(pretty = TRUE)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{trees}}{(list) list of phylogenetic trees}

\item{\code{query}}{(character) query string}

\item{\code{data}}{ignored for now}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{PhyloDiv$print()}}
\item \href{#method-new}{\code{PhyloDiv$new()}}
\item \href{#method-add_tree}{\code{PhyloDiv$add_tree()}}
\item \href{#method-fetch_hierarchies}{\code{PhyloDiv$fetch_hierarchies()}}
\item \href{#method-to_list}{\code{PhyloDiv$to_list()}}
\item \href{#method-to_json}{\code{PhyloDiv$to_json()}}
\item \href{#method-clone}{\code{PhyloDiv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print method for `PhyloDivOne` objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `PhyloDiv` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any number of `PhyloDivOne` objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PhyloDiv` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_tree"></a>}}
\subsection{Method \code{add_tree()}}{
add a tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$add_tree(tree)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing, adds a tree
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch_hierarchies"></a>}}
\subsection{Method \code{fetch_hierarchies()}}{
fetch hierarchy data for all trees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$fetch_hierarchies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\subsection{Method \code{to_list()}}{
coerce to a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of lists
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\subsection{Method \code{to_json()}}{
coerce to JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$to_json(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json/character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhyloDiv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
